+++ tests/test_markets.py
@@ -125,7 +125,7 @@ class MarketTest(unittest.TestCase):
         self.market.place_order(mock_order)
         mock_order.place.assert_called_with(self.market.market_book.publish_time)
         self.assertEqual(self.market.blotter.pending_place, [mock_order])
-        self.mock_flumine.log_control.assert_called_with(mock_events.OrderEvent())
+        self.mock_flumine.log_control.assert_called_with(mock_events.TradeEvent())
 
     @mock.patch("flumine.markets.market.events")
     def test_place_order_not_executed(self, mock_events):
@@ -134,7 +134,7 @@ class MarketTest(unittest.TestCase):
         self.market.place_order(mock_order, execute=False)
         mock_order.place.assert_called_with(self.market.market_book.publish_time)
         self.assertEqual(self.market.blotter.pending_place, [])
-        self.mock_flumine.log_control.assert_called_with(mock_events.OrderEvent())
+        self.mock_flumine.log_control.assert_called_with(mock_events.TradeEvent())
 
     def test_place_order_retry(self):
         mock_order = mock.Mock()